requirements.txt
attrs==22.2.0
jsonschema==4.17.3
pre-commit-hooks==0.1.0
pyrsistent==0.19.3
PyYAML==6.0
